+++ src/environment-handlers/browser.js
@@ -18,9 +18,6 @@ import {Logger} from "../logger.js"
  *   id: string
  * }} CommandsRequireContextType
  */
-/**
- * @typedef {() => Promise<CommandsRequireContextType>} CommandsRequireContextCallbackType
- */
 
 export default class VelociousEnvironmentsHandlerBrowser extends Base {
   /** @type {CommandsRequireContextType | undefined} */
@@ -31,14 +28,12 @@ export default class VelociousEnvironmentsHandlerBrowser extends Base {
 
   /**
    * @param {object} args - Options object.
-   * @param {CommandsRequireContextCallbackType} [args.commandsRequireContextCallback] - Commands require context callback.
    * @param {() => Promise<MigrationsRequireContextType>} [args.migrationsRequireContextCallback] - Migrations require context callback.
    */
-  constructor({commandsRequireContextCallback, migrationsRequireContextCallback, ...restArgs} = {}) {
+  constructor({migrationsRequireContextCallback, ...restArgs} = {}) {
     super()
     restArgsError(restArgs)
 
-    this.commandsRequireContextCallback = commandsRequireContextCallback
     this.migrationsRequireContextCallback = migrationsRequireContextCallback
     this.logger = new Logger(this)
   }
@@ -61,25 +56,23 @@ export default class VelociousEnvironmentsHandlerBrowser extends Base {
    * @returns {Promise<Array<import("./base.js").CommandFileObjectType>>} - Resolves with the commands.
    */
   async findCommands() {
-    this._findCommandsResult = await this._actualFindCommands()
+    this._findCommandsResult = this._actualFindCommands()
 
     return this._findCommandsResult
   }
 
   /**
-   * @returns {Promise<CommandsRequireContextType | undefined>} - The commands require context.
+   * @returns {CommandsRequireContextType} - The commands require context.
    */
-  async _findCommandsRequireContext() {
-    if (!this.commandsRequireContextCallback) return undefined
-
-    this.findCommandsRequireContextResult ||= await this.commandsRequireContextCallback()
+  _findCommandsRequireContext() {
+    // @ts-expect-error
+    this.findCommandsRequireContextResult ||= /** @type {CommandsRequireContextType} */ (require.context("../cli/commands", true, /\.js$/))
 
     return this.findCommandsRequireContextResult
   }
 
-  async _actualFindCommands() {
-    const commandFiles = await this._findCommandsRequireContext()
-    if (!commandFiles) return []
+  _actualFindCommands() {
+    const commandFiles = this._findCommandsRequireContext()
     const commands = []
 
     for (const aFilePath of commandFiles.keys()) {
@@ -122,9 +115,6 @@ export default class VelociousEnvironmentsHandlerBrowser extends Base {
     filePaths.push(filePath)
 
     const commandsRequireContext = await this._findCommandsRequireContext()
-    if (!commandsRequireContext) {
-      throw new Error("CLI commands are not available in the browser environment without commandsRequireContextCallback")
-    }
     let commandClassImport
 
     for (const aFilePath of filePaths) {