+++ src/environment-handlers/node.js
@@ -9,6 +9,9 @@ import CliCommandsGenerateModel from "./node/cli/commands/generate/model.js"
 import CliCommandsRoutes from "./node/cli/commands/routes.js"
 import CliCommandsServer from "./node/cli/commands/server.js"
 import CliCommandsTest from "./node/cli/commands/test.js"
+import CliCommandsBackgroundJobsMain from "./node/cli/commands/background-jobs-main.js"
+import CliCommandsBackgroundJobsWorker from "./node/cli/commands/background-jobs-worker.js"
+import CliCommandsBackgroundJobsRunner from "./node/cli/commands/background-jobs-runner.js"
 import {dirname} from "path"
 import {fileURLToPath} from "url"
 import fs from "fs/promises"
@@ -170,6 +173,30 @@ export default class VelociousEnvironmentHandlerNode extends Base{
    */
   async cliCommandsTest(command) {
     return await this.forwardCommand(command, CliCommandsTest)
+  }
+
+  /**
+   * @param {import("../cli/base-command.js").default} command - Command.
+   * @returns {Promise<unknown>} - Resolves with the command result.
+   */
+  async cliCommandsBackgroundJobsMain(command) {
+    return await this.forwardCommand(command, CliCommandsBackgroundJobsMain)
+  }
+
+  /**
+   * @param {import("../cli/base-command.js").default} command - Command.
+   * @returns {Promise<unknown>} - Resolves with the command result.
+   */
+  async cliCommandsBackgroundJobsWorker(command) {
+    return await this.forwardCommand(command, CliCommandsBackgroundJobsWorker)
+  }
+
+  /**
+   * @param {import("../cli/base-command.js").default} command - Command.
+   * @returns {Promise<unknown>} - Resolves with the command result.
+   */
+  async cliCommandsBackgroundJobsRunner(command) {
+    return await this.forwardCommand(command, CliCommandsBackgroundJobsRunner)
   }
 
   /**